Revert "monet: colors: Switch to single-precision floating-point math"
This reverts commit ed5f03fa99ec477c5386da25acf1b17a039605fb.

There are still small differences in pure blue (#0000FF) output,
especially with adaptive L0 towards L_cusp clipping, and double
performance is good enough for our needs.

+++ app/src/main/java/dev/kdrag0n/android12ext/core/xposed/XposedHook.kt
@@ -78,7 +78,7 @@ class XposedHook(
 
             sysuiHooks.applyThemeOverlayController(
                 isGoogle = hasSystemUiGoogle,
-                chromaMultiplier = prefs.getInt("custom_monet_chroma_multiplier", 50).toFloat() / 50,
+                chromaMultiplier = prefs.getInt("custom_monet_chroma_multiplier", 50).toDouble() / 50,
                 multiColor = false,
                 accurateShades = isFeatureEnabled("custom_monet_accurate_shades"),
                 colorOverride = colorOverride,
